Added methods to Half Edge Mesh classes
A number of methods were added to the Half Edge Mesh classes to increase functionality and allow deletions/modifications of the data in these classes.
All changes should be backwards compatible and not effect existing functions.
Of particular, note:
VertexBase: Method IsBoundary() returns true if the vertex is on the boundary.
HalfEdgeBase: Method MinIndexHalfEdgeAroundEdge() returns the minimum index of the half edge around the same edge as the current half edge. Useful if you are trying to find a unique half edge for each edge.
HalfEdgeMeshBase: Multiple methods for counting number of vertices, edges, cells of different types, for finding a specific edge in the mesh, and for deleting vertices, edges and cells.